// src/models/loby/loby.model.ts
export interface Loby {
  id: number;
  curp: number;
  nombre: string;

  servicios: string;
  edad: number;
  telefono: number;
  codigo_postal: number;
  colonia: string;
  estado: string;
  municipio: string;
  localidad: string;
  calle: string;
  num_ext: string;
  num_int: string;
  entre_calles: string;
  referencias: string;
  zona: string;
}